Use UTC datetimes internally
Python datetime object comparisons don't work as expected when the
objects are aware and have the same tzinfo attribute. Basically the fold
attribute is ignored in this case so comparing times before and after a
DST to standard time switch returns the wrong result. Avoid this problem
by using aware times in UTC internally instead of the local tz. Only use
the local tz for user visible attributes.